refactor: einkvnc as library
